+++ ackpine-splits/src/main/kotlin/ru/solrudev/ackpine/helpers/LocaleHelpers.kt
@@ -25,7 +25,7 @@ import java.util.Locale
 internal fun deviceLocales(context: Context): List<Locale> {
 	val locales = ConfigurationCompat.getLocales(context.resources.configuration)
 	return buildList {
-		for (index in 0 until locales.size()) {
+		for (index in 0..<locales.size()) {
 			add(locales[index]!!)
 		}
 	}